Adding pytest boilerplate
