+++ python/setup.cfg
@@ -1,0 +1,5 @@
+[aliases]
+test=pytest
+
+[flake8]
+max-line-length = 120
